+++ CHANGELOG.md
@@ -2,6 +2,8 @@ Next
 ====
 - Feature: Add PlainNumberFormat that does not round or use scientific notations for long numbers.
            Can be enabled by setting `usePlainNumberFormat=true` when reading the excel file. 
+- Bugfix: Fixed SaveMode.Overwrite and SaveMode.Append for V2 API and spark >=3.0.1 
+- Feature: Writing partitioned file structure for V2 API and spark >=3.0.1
 
 0.13.2
 ======